+++ core/src/main/java/fi/vincit/multiusertest/rule/EmptyUserDefinitionClass.java
@@ -1,0 +1,17 @@
+package fi.vincit.multiusertest.rule;
+
+public class EmptyUserDefinitionClass implements UserDefinitionClass {
+
+    private static final String[] EMPTY_DEFINITIONS = new String[0];
+
+    private static final UserDefinitionClass EMPTY_DEFINITION_CLASS = new EmptyUserDefinitionClass();
+
+    public static UserDefinitionClass getEmptyClass() {
+        return EMPTY_DEFINITION_CLASS;
+    }
+
+    @Override
+    public String[] getUsers() {
+        return EMPTY_DEFINITIONS;
+    }
+}