+++ core/src/main/java/fi/vincit/multiusertest/rule/UserDefinitionClass.java
@@ -1,0 +1,10 @@
+package fi.vincit.multiusertest.rule;
+
+/**
+ * Defines
+ */
+public interface UserDefinitionClass {
+
+    String[] getUsers();
+
+}