+++ examples/src/test/java/fi/vincit/mutrproject/feature/todo/TodoProducers.java
@@ -1,0 +1,13 @@
+package fi.vincit.mutrproject.feature.todo;
+
+import fi.vincit.multiusertest.rule.UserDefinitionClass;
+
+public class TodoProducers implements UserDefinitionClass {
+
+    @Override
+    public String[] getUsers() {
+        return new String[] {
+                "role:ROLE_SYSTEM_ADMIN", "role:ROLE_ADMIN", "role:ROLE_USER", "role:ROLE_USER"
+        };
+    }
+}